Added a timestamp column to database
The timestamp column can be used to track when rows are added or measure elapsed time between insertions. 
Also added a listing after processing all downloads of the hashes for the grupe index files and the hash for the sqlite database which is also now inserted into IPFS after the grupe index file listing.

Lastly there were several bug fixes and minor changes to the code and comments.